+++ startup/gui/menus.py
@@ -371,6 +371,7 @@ nodeMenu.append( "/Scene/Utility/Shader Query", GafferScene.ShaderQuery, searchT
 nodeMenu.append( "/Scene/Utility/Option Query", GafferScene.OptionQuery, searchText = "OptionQuery" )
 nodeMenu.append( "/Scene/Utility/Primitive Query", GafferScene.PrimitiveQuery, searchText = "PrimitiveQuery" )
 nodeMenu.append( "/Scene/Utility/Primitive Variable Query", GafferScene.PrimitiveVariableQuery, searchText = "PrimitiveVariableQuery" )
+nodeMenu.append( "/Scene/Utility/Point Instancer Query", GafferScene.PointInstancerQuery, searchText = "PointInstancerQuery" )
 nodeMenu.append( "/Scene/Utility/Camera Query", GafferScene.CameraQuery, searchText = "CameraQuery" )
 nodeMenu.append( "/Scene/Utility/Scene Stats", GafferScene.SceneStats, searchText = "SceneStats" )
 nodeMenu.append( "/Scene/Passes/Render Passes", GafferScene.RenderPasses, searchText = "RenderPasses" )