+++ Changes.md
@@ -1,6 +1,11 @@
 1.7.x.x (relative to 1.7.2.0)
 =======
 
+Features
+--------
+
+- PointInstancerQuery : Added a new node for querying the properties (prototype, transform, visibility etc) of a specific instance within a PointInstancer.
+
 Improvements
 ------------